OPA_APP/stream/requirements.txt
pip
websocket-client
pandas
python-binance
rel
python-dateutil
psycopg2-binary